requirements.txt
alembic==1.16.1
attrs==25.3.0
bleach==6.2.0
blinker==1.9.0
boto3==1.38.37
botocore==1.38.37
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
docutils==0.21.2
EbookLib==0.19
filelock==3.0.12
Flask==3.1.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
FlaskSQL==1.0.3
idna==2.8
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lockfile==0.12.2
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyMySQL==1.1.1
python-daemon==2.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
referencing @ git+https://github.com/python-jsonschema/referencing.git@a4e5485c92266ad74eddf72406f0e6fba2c9fe3a
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
s3transfer==0.13.0
setuptools==80.8.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
